requirements.txt
Django==5.2.6
djangorestframework==3.16.1
django-cors-headers==4.9.0
django-filter==23.5
psycopg2-binary==2.9.9
Pillow==10.1.0
asgiref==3.9.2
sqlparse==0.5.3
